requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
httpie==3.2.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
multidict==6.0.2
Pygments==2.13.0
PyJWT==2.5.0
pymongo==4.2.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
